Checking DB working, and extraction feature being added

+++ src/database_setup.py
@@ -1,6 +1,6 @@
 import os
 import sys
-from sqlalchemy import Column, ForeignKey, Integer, String, DateTime
+from sqlalchemy import Column, ForeignKey, Integer, String, DateTime, Float
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 from sqlalchemy import create_engine
@@ -13,7 +13,7 @@ class Clickbait(Base):
 
     id = Column(Integer, primary_key=True)
     title = Column(String(500), nullable=False)
-    clickbaitiness = Column(Integer, nullable=False)
+    clickbaitiness = Column(Float, nullable=False)
 
     @property
     def serialize(self):
@@ -32,14 +32,14 @@ class Stance(Base):
     body   = Column(String(1500), nullable=False)
     stance = Column(String(250))
 
-     @property
-     def serialize(self):
-         """Return object data in easily serializeable format"""
-         return {
-             'id': self.id,
-             'title': self.title,
-             'stance': self.stance
-         }
+    @property
+    def serialize(self):
+        """Return object data in easily serializeable format"""
+        return {
+            'id': self.id,
+            'title': self.title,
+            'stance': self.stance
+        }
 
 
 engine = create_engine('sqlite:///cnn.db')
